Refactor: Move lottery ball & empty lottery ball to components package
