+++ app/src/main/java/com/bonustrack02/lotterygenerator/activities/MainActivity.kt
@@ -4,30 +4,22 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.NavigationBar
 import androidx.compose.material3.NavigationBarItem
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBar
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.navigation.activity
 import androidx.navigation.compose.NavHost
@@ -43,12 +35,7 @@ import com.bonustrack02.lotterygenerator.presentation.navigation.Screen
 import com.bonustrack02.lotterygenerator.presentation.navigation.bottomNavItems
 import com.bonustrack02.lotterygenerator.presentation.settings.SettingsScreen
 import com.bonustrack02.lotterygenerator.ui.components.AdmobBanner
-import com.bonustrack02.lotterygenerator.ui.theme.LotteryBlue
 import com.bonustrack02.lotterygenerator.ui.theme.LotteryGeneratorTheme
-import com.bonustrack02.lotterygenerator.ui.theme.LotteryGray
-import com.bonustrack02.lotterygenerator.ui.theme.LotteryGreen
-import com.bonustrack02.lotterygenerator.ui.theme.LotteryRed
-import com.bonustrack02.lotterygenerator.ui.theme.LotteryYellow
 import com.google.android.gms.ads.AdSize
 import dagger.hilt.android.AndroidEntryPoint
 
@@ -140,46 +127,4 @@ class MainActivity : ComponentActivity() {
             }
         }
     }
-}
-
-@Composable
-fun LotteryBall(
-    number: Int,
-    modifier: Modifier = Modifier.Companion
-) {
-    val ballColor = when (number) {
-        in 1..10 -> LotteryYellow
-        in 11..20 -> LotteryBlue
-        in 21..30 -> LotteryRed
-        in 31..40 -> LotteryGray
-        else -> LotteryGreen
-    }
-
-    Box(
-        contentAlignment = Alignment.Center,
-        modifier = modifier
-            .size(40.dp)
-            .background(color = ballColor, shape = CircleShape)
-    ) {
-        Text(
-            text = number.toString(),
-            color = Color.White,
-            fontWeight = FontWeight.Bold,
-            style = MaterialTheme.typography.bodyMedium
-        )
-    }
-}
-
-@Composable
-fun EmptyLotteryBall(
-    modifier: Modifier = Modifier.Companion
-) {
-    Box(
-        modifier = modifier
-            .size(40.dp) // LotteryBall과 동일한 크기 유지
-            .background(
-                color = Color.LightGray.copy(alpha = 0.3f),
-                shape = CircleShape
-            )
-    )
 }
